REFACTOR: Update database connection to export Prisma client and remove SMTP config; enhance error handling in controllers and create middleware to verify body or query params fields.

+++ src/service/UserService.js
@@ -1,7 +1,6 @@
-import { PrismaClient } from "@prisma/client";
+import { ObjectId } from "mongodb";
 import bcrypt from "bcrypt";
-
-const prisma = new PrismaClient();
+import { prisma } from "../config/connectDb.js";
 
 export default class UserService {
   static async createUser(payload = {}) {
@@ -11,14 +10,14 @@ export default class UserService {
       throw new Error("Email já cadastrado");
     }
 
-    const formattedDate = new Date(birthday + "T00:00:00.000Z");
+    const formattedDate = new Date(`${birthday}T00:00:00Z`).toISOString();
     const hashedPassword = await bcrypt.hash(password, 10);
 
     const user = await prisma.User.create({
       data: {
         name,
         email,
-        birthday,
+        birthday: formattedDate,
         password: hashedPassword,
       },
     });
@@ -27,7 +26,7 @@ export default class UserService {
       id: user.id,
       name: user.name,
       email: user.email,
-      birthday: user.birthday,
+      birthday: formattedDate,
     };
   }
 
@@ -54,14 +53,33 @@ export default class UserService {
     });
   }
 
-  static async updateUser(userId, data) {
+  static async updateUser(userId, payload = {}) {
+    if (!ObjectId.isValid(userId)) {
+      throw new Error("ID inválido: o ID fornecido não é um ObjectId válido.");
+    }
+
+    if (!userId) {
+      throw new Error("ID Inválido: usuário não encontrado.");
+    }
+
     return prisma.User.update({
       where: { id: userId },
-      data,
+      payload,
     });
   }
 
   static async deleteUser(userId) {
+    const userExists = await prisma.User.findUnique({
+      where: { id: userId },
+    });
+
+    if (!ObjectId.isValid(userId)) {
+      throw new Error("ID inválido: o ID fornecido não é um ObjectId válido.");
+    }
+    if (!userExists) {
+      throw new Error("ID Inválido: usuário não encontrado.");
+    }
+
     return prisma.User.delete({
       where: { id: userId },
     });
